ingest/requirements.txt
pandas==2.2.0
psycopg2-binary==2.9.9
requests==2.32.0
pytest==8.3.0
